refactor: split UserdefaultSetting responsibilities into stores
